Refactor to use :projector-util-loading
