+++ projector-agent/src/main/kotlin/org/jetbrains/projector/agent/GraphicsInterceptor.kt
@@ -30,7 +30,7 @@ import org.jetbrains.projector.awt.peer.PMouseInfoPeer
 import org.jetbrains.projector.common.protocol.toClient.ServerDrawCommandsEvent
 import org.jetbrains.projector.common.protocol.toClient.ServerWindowEvent
 import org.jetbrains.projector.server.ProjectorServer
-import org.jetbrains.projector.server.core.util.unprotect
+import org.jetbrains.projector.util.loading.unprotect
 import org.jetbrains.projector.server.service.ProjectorDrawEventQueue
 import org.jetbrains.projector.server.service.ProjectorFontProvider
 import org.jetbrains.projector.util.logging.Logger
