+++ projector-server/src/main/kotlin/org/jetbrains/projector/server/HeadlessInitializer.kt
@@ -29,7 +29,7 @@ import org.jetbrains.projector.awt.PToolkit
 import org.jetbrains.projector.awt.font.PFontManager
 import org.jetbrains.projector.awt.image.PGraphicsEnvironment
 import org.jetbrains.projector.server.ProjectorServer.Companion.ENABLE_PROPERTY_NAME
-import org.jetbrains.projector.server.core.util.unprotect
+import org.jetbrains.projector.util.loading.unprotect
 import sun.font.FontManagerFactory
 import java.awt.GraphicsEnvironment
 import java.awt.Toolkit
