+++ projector-server/src/main/kotlin/org/jetbrains/projector/server/idea/CaretInfoUpdater.kt
@@ -32,8 +32,8 @@ import org.jetbrains.projector.common.protocol.data.Point
 import org.jetbrains.projector.common.protocol.toClient.ServerCaretInfoChangedEvent
 import org.jetbrains.projector.common.protocol.toClient.data.idea.CaretInfo
 import org.jetbrains.projector.server.core.ij.invokeWhenIdeaIsInitialized
-import org.jetbrains.projector.server.core.util.unprotect
 import org.jetbrains.projector.server.util.FontCacher
+import org.jetbrains.projector.util.loading.unprotect
 import org.jetbrains.projector.util.logging.Logger
 import sun.awt.AWTAccessor
 import java.awt.Component
