+++ projector-server/src/main/kotlin/org/jetbrains/projector/server/util/FontCacher.kt
@@ -28,8 +28,8 @@ package org.jetbrains.projector.server.util
 import org.jetbrains.projector.common.protocol.data.FontDataHolder
 import org.jetbrains.projector.common.protocol.data.TtfFontData
 import org.jetbrains.projector.server.core.util.ObjectIdCacher
-import org.jetbrains.projector.server.core.util.unprotect
 import org.jetbrains.projector.server.service.ProjectorFontProvider
+import org.jetbrains.projector.util.loading.unprotect
 import sun.font.CompositeFont
 import sun.font.FileFont
 import sun.font.Font2D
